+++ camera.h
@@ -8,9 +8,9 @@ class Camera
 public:
     // TODO: probably requires a builder class
     Camera(); // orients camera facing into the screen at the origin
-    Camera(Math::Vector3 viewPoint, Math::Vector3 u, Math::Vector3 v, Math::Vector3 w, int resolutionX, int resolutionY, float leftBound, float rightBound, float topBound, float bottomBound);
+    Camera(Math::Vector3 viewPoint, Math::Vector3 u, Math::Vector3 v, Math::Vector3 w, int resolutionX, int resolutionY, double leftBound, double rightBound, double topBound, double bottomBound);
     // places camera at viewPoint and orients it facing the viewing direction with v pointing up
-    Camera(Math::Vector3 viewPoint, Math::Vector3 viewingDirection, int resolutionX, int resolutionY, float leftBound, float rightBound, float topBound, float bottomBound);
+    Camera(Math::Vector3 viewPoint, Math::Vector3 viewingDirection, int resolutionX, int resolutionY, double leftBound, double rightBound, double topBound, double bottomBound);
 
     int getResolutionX() const;
     int getResolutionY() const;
@@ -19,7 +19,7 @@ public:
     void setResolution(int resolutionX, int resolutionY);
     void setOrientation(Math::Vector3 u, Math::Vector3 v, Math::Vector3 w);
     void setOrientation(Math::Vector3 viewingDirection); // points the camera toward the viewing direction
-    void setBounds(float leftBound, float rightBound, float topBound, float bottomBound);
+    void setBounds(double leftBound, double rightBound, double topBound, double bottomBound);
 
     virtual Math::Ray computeViewingRay(int pixelIndexX, int pixelIndexY) const = 0;
 
@@ -29,7 +29,7 @@ protected:
     // orthonormal basis of the camera
     // up points right, v upwards, and w against the direction of the camera
     Math:: Vector3 u, v, w;
-    float leftBound, rightBound, topBound, bottomBound; // labeled l, r, u, d in the text
+    double leftBound, rightBound, topBound, bottomBound; // labeled l, r, u, d in the text
 };
 
 class ParallelOrthographicCamera: public Camera
@@ -42,15 +42,15 @@ class PerspectiveCamera: public Camera
 {
 public:
     PerspectiveCamera();
-    PerspectiveCamera(Math::Vector3 viewPoint, Math::Vector3 u, Math::Vector3 v, Math::Vector3 w, int resolutionX, int resolutionY, float leftBound, float rightBound, float topBound, float bottomBound, float focalLength);
-    PerspectiveCamera(Math::Vector3 viewPoint, Math::Vector3 viewingDirection, int resolutionX, int resolutionY, float leftBound, float rightBound, float topBound, float bottomBound, float focalLength);
+    PerspectiveCamera(Math::Vector3 viewPoint, Math::Vector3 u, Math::Vector3 v, Math::Vector3 w, int resolutionX, int resolutionY, double leftBound, double rightBound, double topBound, double bottomBound, double focalLength);
+    PerspectiveCamera(Math::Vector3 viewPoint, Math::Vector3 viewingDirection, int resolutionX, int resolutionY, double leftBound, double rightBound, double topBound, double bottomBound, double focalLength);
 
-    void setFocalLength(float focalLength);
+    void setFocalLength(double focalLength);
 
     Math::Ray computeViewingRay(int pixelIndexX, int pixelIndexY) const;
 
 protected:
-    float focalLength = 0;
+    double focalLength = 0;
 };
 
 #endif
